refactor(lib): break import cycles and enforce no-cycle in lint
madge reported 23 circular chains on the base branch; all are now broken
with zero public-surface and zero behavior change (audit roadmap 4.1.6,
stacked on the storage type-hygiene branch, PR #530):

- Facade back-imports (cycles 2-14, 16-22): every lib/storage/* submodule
  that imported types back from "../storage.js" now imports them from the
  layer below — AccountStorageV3/AccountMetadataV3/FlaggedAccountStorageV1
  from "./public-types.js", BackupMetadata/RestoreAssessment from
  "./backup-metadata.js" (definitions moved there verbatim next to their
  constituent section/snapshot types), NamedBackupSummary from its defining
  module "./named-backups.js". FlaggedAccountMetadataV1/FlaggedAccountStorageV1
  moved verbatim into storage/public-types.ts; lib/storage.ts re-exports
  everything it exported before.
- schemas cycle (1): ModelFamily/MODEL_FAMILIES moved verbatim from
  lib/prompts/codex.ts to the leaf module lib/request/helpers/model-map.ts
  (which already defined PromptModelFamily); prompts/codex.ts re-exports
  them, schemas.ts and storage/public-types.ts import the leaf directly.
- accounts -> codex-cli -> storage cycle (15): Workspace moved verbatim
  from lib/accounts.ts into storage/public-types.ts (it is part of the
  persisted storage shape) with a re-export from accounts.ts, and
  codex-cli/sync.ts retargets its type-only AccountStorageV3 import to
  storage/public-types.ts.
- accounts -> codex-manager -> forecast cycle (23): the shared VALUE
  helpers saveAccountsWithRetry/isRetryableStorageWriteError moved verbatim
  to a new lib/storage/save-retry.ts; forecast-report-shared.ts re-exports
  them for its codex-manager consumers and accounts.ts imports the lower
  module.

Enforcement: eslint-plugin-import-x (flat-config native, ESLint 10
compatible) + eslint-import-resolver-typescript added as devDependencies;
"import-x/no-cycle": ["error", { maxDepth: Infinity, ignoreExternal: true }]
scoped to lib/** and index.ts only (scripts/ and test/ stay out of scope).
"import-x/extensions": [".ts"] is required — without it import-x silently
skips .ts files during cycle traversal. Verified the rule actually fires by
temporarily re-adding a lib/storage/transactions.ts -> ../storage.js import
(eslint reported import-x/no-cycle) and then removing it.

Documented the dependency direction (types/constants -> storage ->
accounts -> runtime -> manager/CLI) in lib/AGENTS.md conventions.

Verification: madge --circular lib/ index.ts reports 0 cycles; typecheck
and full lint pass; storage/accounts/codex-cli suites show the identical
41 known environment-only EACCES failures as the base branch (555 passed
on both). Lockfile diff is additions-only for the two new dev packages
(rollup optional-dep "libc" entries restored byte-identical).
