+++ lib/codex-cli/sync.ts
@@ -1,6 +1,6 @@
 import { createLogger } from "../logger.js";
 import { MODEL_FAMILIES, type ModelFamily } from "../prompts/codex.js";
-import { type AccountStorageV3 } from "../storage.js";
+import { type AccountStorageV3 } from "../storage/public-types.js";
 import { incrementCodexCliMetric } from "./observability.js";
 
 const log = createLogger("codex-cli-sync");
