+++ lib/storage/account-save-entry.ts
@@ -1,4 +1,4 @@
-import type { AccountStorageV3 } from "../storage.js";
+import type { AccountStorageV3 } from "./public-types.js";
 
 export async function saveAccountsEntry(params: {
 	storage: AccountStorageV3;
