+++ lib/storage/backup-restore.ts
@@ -1,5 +1,5 @@
 import { isAbsolute, relative } from "node:path";
-import type { AccountStorageV3 } from "../storage.js";
+import type { AccountStorageV3 } from "./public-types.js";
 
 export async function restoreAccountsFromBackupPath(
 	path: string,
